Try to perform safe copy of properties only if explicitly specified on the exchange
